requirements.txt
altair==5.5.0
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
gitdb==4.0.12
gitpython==3.1.45
idna==3.10
importlib-metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jinja2==3.1.6
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
markupsafe==3.0.2
more-itertools==10.3.0
narwhals==2.5.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
platformdirs==4.2.2
protobuf==6.32.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
schedule==1.2.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.0.1
tornado==6.5.2
typeguard==4.3.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wheel==0.45.1
zipp==3.19.2